Revisiting import asset process (#449)
* revisited import asset process

* revisited asset manager

* removed dead codes

* fixed method signature

* removed static qualifiers

* fixed build errors

* added node hierarchy uuid

* added noderef support for editorscene hierarchy

* fixed assetManager shutdown

* fixed assetManager shutdown proc

+++ ZEngine/ZEngine/Core/Memory/MemoryManager.h
@@ -5,7 +5,7 @@ namespace ZEngine::Core::Memory
 {
     struct MemoryConfiguration
     {
-        size_t DefaultSize = ZGiga(2ull);
+        uint64_t DefaultSize = ZGiga(2ull);
     };
 
     struct MemoryManager
